For server

+++ main.py
@@ -19,14 +19,15 @@ def monitor_delay_messages():
     while True:
         time.sleep(5)
         row_msg = db.get_early_delay_message()
-        if len(row_msg) != 0:
-            date_time_msg = datetime.datetime.strptime(row_msg[0][5],"%Y-%m-%d %H:%M:%S")
-            date_time_now = datetime.datetime.now() + timedelta(hours=int(globals.DELTA_TIME_SERVER))
-            delta = (date_time_msg - date_time_now).total_seconds()
+        if row_msg != None:
+            if len(row_msg) != 0:
+                date_time_msg = datetime.datetime.strptime(row_msg[0][5],"%Y-%m-%d %H:%M:%S")
+                date_time_now = datetime.datetime.now() + timedelta(hours=int(globals.DELTA_TIME_SERVER))
+                delta = (date_time_msg - date_time_now).total_seconds()
             
-            if delta <= 0:
-                mailing(row_msg[0][1], row_msg[0][2], row_msg[0][3], db=db)
-                db.delete_message_for_it_id(row_msg[0][0])
+                if delta <= 0:
+                    mailing(row_msg[0][1], row_msg[0][2], row_msg[0][3], db=db)
+                    db.delete_message_for_it_id(row_msg[0][0])
 
 def start_polling():
     inited = init()
